+++ src/ts/lib/01-config/index.ts
@@ -8,6 +8,4 @@
  * @license MIT
  */
 
-export * from './classes/index.js';
-
 export * from './parseParamsCLI.js';