+++ src/ts/lib/03-stages/classes/SnapshotStage.ts
@@ -23,6 +23,7 @@ import {
 
 import type {
     CLI,
+    Config,
     Stage,
 } from '../../../types/index.js';
 
@@ -34,9 +35,8 @@ import {
     FileSystem,
 } from '../../00-universal/index.js';
 
-import {
-    ProjectConfig,
-} from '../../01-config/index.js';
+// import {
+// } from '../../01-config/index.js';
 
 import { AbstractStage } from './abstract/AbstractStage.js';
 
@@ -109,7 +109,7 @@ export class SnapshotStage extends AbstractStage<
      * @param version  Version object for the project’s version.
      */
     constructor (
-        config: ProjectConfig,
+        config: Config.Class,
         params: CLI.Params,
         args: Partial<Stage.Args.Snapshot>,
         pkg?: Json.PackageJson,