+++ src/ts/lib/04-project/classes/Project.ts
@@ -20,7 +20,6 @@ import { DummyConsole } from '../../@internal/index.js';
 
 import {
     parseParamsCLI,
-    ProjectConfig,
 } from '../../01-config/index.js';
 
 // import {
@@ -31,6 +30,8 @@ import { Stage_Console } from '../../02-utils/classes/Stage_Console.js';
 import {
     defaultConfig,
 } from '../../03-stages/defaultConfig.js';
+
+import { ProjectConfig } from './ProjectConfig.js';
 
 /**
  * Manages and runs a single project (typically used by the cli).