+++ src/ts/lib/04-project/classes/index.ts
@@ -9,3 +9,4 @@
  */
 
 export * from './Project.js';
+export * from './ProjectConfig.js';