+++ src/ts/lib/@internal.ts
@@ -33,6 +33,7 @@ export * from './@internal/classes/SemVer.js';
 
 export * from './@internal/errorHandler.js';
 export * from './@internal/errorStringify.js';
+export * from './@internal/isObjectEmpty.js';
 export * from './@internal/logError.js';
 export * from './@internal/writeLog.js';